requirements.txt
numpy
pandas
matplotlib
seaborn
requests
opencv-python
tqdm
lxml
plotly
ipywidgets
ipykernel
jupyterlab>=3.0.0
setuptools
sklearn
comet_ml
python-dotenv